consolidate plugins
